🛠 rename Feature QRCode to QRScan(#5)
