+++ app/src/main/java/com/mashup/ui/mypage/MyPageFragment.kt
@@ -3,7 +3,6 @@ package com.mashup.ui.mypage
 import com.mashup.R
 import com.mashup.base.BaseFragment
 import com.mashup.databinding.FragmentMyPageBinding
-import com.mashup.ui.qrcode.QRCodeFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
