+++ app/src/main/java/com/mashup/ui/qrscan/camera/BaseImageAnalyzer.kt
@@ -1,4 +1,4 @@
-package com.mashup.ui.qrcode.camera
+package com.mashup.ui.qrscan.camera
 
 import android.annotation.SuppressLint
 import android.graphics.Rect
