+++ app/src/main/res/layout/fragment_qr_scan.xml
